#12 création du multi module.
