+++ .env
@@ -1,0 +1,4 @@
+FORTRESS_FINANCE_PROVIDER_EXPENSE_MONGO_URL="mongodb://192.168.49.2:30017"
+FORTRESS_FINANCE_PROVIDER_EXPENSE_MONGO_USER="fortress"
+FORTRESS_FINANCE_PROVIDER_EXPENSE_MONGO_PASSWORD="fortress_password"
+FORTRESS_FINANCE_PROVIDER_EXPENSE_MONGO_DATABASE="admin"
