--- bootstrap/src/main/java/me/noynto/fortress/finance/Provider.java
@@ -1,12 +1,0 @@
-package me.noynto.fortress.finance;
-
-public interface Provider {
-    class NotAvailable extends RuntimeException {
-        public NotAvailable(String message) {
-            super(message);
-        }
-        public NotAvailable(String message, Throwable cause) {
-            super(message, cause);
-        }
-    }
-}
