+++ bootstrap/src/test/java/me/noynto/fortress/bootstrap/BootstrapTest.java
@@ -1,0 +1,12 @@
+package me.noynto.fortress.bootstrap;
+
+
+import org.junit.jupiter.api.Test;
+
+class BootstrapTest {
+
+    @Test
+    void testItWorks() {
+    }
+
+}
