--- bootstrap/src/test/java/me/noynto/fortress/finance/BootstrapTest.java
@@ -1,15 +1,0 @@
-package me.noynto.fortress.finance;
-
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class BootstrapTest {
-
-    @Test
-    void appHasAGreeting() {
-        Bootstrap.main(null);
-    }
-
-}
