+++ finance-application/pom.xml
@@ -1,0 +1,18 @@
+<project>
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>me.noynto.fortress</groupId>
+        <artifactId>fortress</artifactId>
+        <version>0.0.0</version>
+    </parent>
+
+    <artifactId>finance-application</artifactId>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+            <version>2.0.9</version>
+        </dependency>
+    </dependencies>
+</project>
