+++ finance-application/src/main/java/me/noynto/fortress/Action.java
@@ -1,0 +1,11 @@
+package me.noynto.fortress;
+
+public interface Action<A> {
+    A execute() throws Failed;
+
+    class Failed extends Exception {
+        public Failed(String message) {
+            super(message);
+        }
+    }
+}
