+++ finance-application/src/main/java/me/noynto/fortress/Provider.java
@@ -1,0 +1,12 @@
+package me.noynto.fortress;
+
+public interface Provider {
+    class NotAvailable extends RuntimeException {
+        public NotAvailable(String message) {
+            super(message);
+        }
+        public NotAvailable(String message, Throwable cause) {
+            super(message, cause);
+        }
+    }
+}
