+++ finance-infrastructure/src/main/java/me/noynto/fortress/finance/infrastructure/SpiFactory.java
@@ -1,0 +1,9 @@
+package me.noynto.fortress.finance.infrastructure;
+
+public class SpiFactory {
+
+    /*public ExpenseProvider expenseProvider() {
+        return new MongoExpense();
+    }*/
+
+}
