--- gradle/libs.versions.toml
@@ -1,12 +1,0 @@
-# https://docs.gradle.org/current/userguide/platforms.html#sub::toml-dependencies-format
-
-[versions]
-junit = "6.0.0"
-mockito = "5.20.0"
-jackson = "3.0.2"
-
-[libraries]
-junit-jupiter = { group = "org.junit.jupiter", name = "junit-jupiter", version.ref = "junit" }
-junit-platform = { group = "org.junit.platform", name = "junit-platform-launcher", version.ref = "junit" }
-mockito = { group = "org.mockito", name = "mockito-core", version.ref = "mockito" }
-jackson-databind = { group = "tools.jackson.core", name = "jackson-databind", version.ref = "jackson" }
